BREAKING CHANGE: Current variable setting now defaults to --pass-by-ref
Objects, arrays, and strings (sine a few commits ago), are set by
passing in the name of the variable that holds either the object, array,
or string. bash-object would copy that info into the new location
correctly. Now, we force the consumer to pass --pass-by-ref due to the
introduction of the --pass-by-value option. --pass-by-value has not been
implemented yet