feat: Existing strings are overridable on presence of var
If `BASH_OBJECT_FORCE_SET` is `yes`, then do not print an error if the
user is attempting to set a non-string to a string (within objects an
arrays). Instead, override the string with that object or string.

Later, this will be implemented for objects and arrays, but only strings
are implemented for now since the former two involved careful
deallocation.